# PERF: make RangeIndex iterate over ._range