--- a/pandas/tests/indexes/ranges/test_range.py
+++ b/pandas/tests/indexes/ranges/test_range.py
@@ -167,6 +167,10 @@ class TestRangeIndex(Numeric):
         idx.any()
         assert idx._cache == {}
 
+        for _ in idx:
+            ...
+        assert idx._cache == {}
+
         df = pd.DataFrame({"a": range(10)}, index=idx)
 
         df.loc[50]